backend/backend_v2/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1

# ---- numerical stack (PyPortfolioOpt requires numpy < 2.0) ----
numpy==1.26.4
pandas==2.2.3
scipy==1.14.1

# ---- finance/data ----
yfinance==0.2.43
statsmodels==0.14.2
arch==7.2.0

# ---- ML + portfolio ----
scikit-learn==1.5.2
PyPortfolioOpt==1.5.5
